Implement last seen label
The ContactsManager now listens to the SocketHandler and notifies it's
listeners (the ContactsTableViewController) in case the last_seen state
of one of the contacts changes.
This was achieved by implementing a decodable enum with cases '.online'
and '.offline' that decodes itself from a string containing either 'online'
or a ISO 8601 encoded date. Sometimes, Swift is somewhat awesome.