Fix Bug where app would crash when adding contacts
Fixed a bug where the app would crash as soon as the table view
showing the search results contains at least one row.
The crash was caused by outlet ids that had not been updated
when the corresponding attributes in the TableViewCell were renamed.

Now also showing the id of possible contacts (instead of email),
because the email is not sent with the response by the backend.